#471 - Move alembic code back to base.py as it makes alembic system test fail
